Eliminate ALSA max channel detection.
This commit does two things:

Eliminates the ALSA max channel detection. I think this made things
more complicated than necessary. This will just use CPAL's
definitions, which may not be 100% accurate, but coupled with the next
change should make things okay.

The song source no longer uses the hardware max channels and instead
just uses the number of channels defined in the song.

The hope here is that this will prevent the mtrack player from detecting
the incorrect song order. I'm not 100% convinced that this is actually a
problem, but this will at least reduce complexity of the code overall.