Use a more accurate MIDI timer, move play barrier. (#32)
The MIDI timer is now more accurate by using the intended next elapsed
time based on the last instant. This helps stop drift in between the
playing of MIDI notes. Additionally, the play barrier has been moved
closer to the MIDI and audio play, which ensures that the audio and MIDI
are closer aligned during playback.