Initial DMX engine.
The initial DMX engine has been implemented. This is not
ready for total use and is missing some key features, but is
capable of playing light shows defined by MIDI on/off notes.

Additionally, the MIDI timer has been fixed, as whatever I had
before was not doing what I thought it was.