feat: DMX playback delay.
A configurable delay has been added to DMX playback. I've found that the
lights play ahead of the click. This will allow us to tune this so that
it lines up more properly with the click.

This should also be implemented for audio playback and MIDI playback,
but those will be separate PRs.