feat: Add gRPC server to the player.
A gRPC server has been added to the player. This will make developing
further controllers a lot easier, as it will no longer be necessary to
explicitly share references to the player between them. Furthermore, it
will make developing external controls unrelated to the mtrack code base
a lot easier as welle.