feat: Switch from serde_yaml to config-rs. (#68)
Instead of using serde_yaml, mtrack now uses config-rs. serde_yaml has
been deprecated, which made the change necessary. config-rs is more
flexible for different file formats, but as a result we've lost support
for multiple YAML song definitions in one file. This was of questionable
use, so I view the hit as well worth the maintenance improvement.