chore: Expand DMX engine testability.
The DMX engine has increased testability. This is largely just
concerning that the DMX engine starts/stops and whether connections
work, but a refactor has taken place to eliminate the need for the
RwLock for the DMX engine by properly scoping "self" in several of the
DMX engine's internal methods.