feat: Transcoding input sources.
Input sources are no longer required to share the same sample rate and
input format. Additionally, the audio format can now be specified as
part of the audio configuration.

The general hope here is to enable an eventual sampling functionality
that supports a steady output thread without needing to reopen or
rebuild the audio output every time.