feat: Light show validation, audio output fixes. (#117)
Light show validation has been added into the playback path, and validation has
been added as a top level command.

Various fixes have been made to the lighting effects engine to make it behave in
the way that I expect it to.

Additionally, a fairly significant audio output bug was discovered in the
transcoder. This has been fixed. Along with it, a move back, to Sinc
interpolation for lower latency has been done.