feat: Songs play from a given time. (#125)
To aide in creation of lighting shows, it's now possible to play a song
from the middle of a song. This will make it easier to test pieces of a
song rather than have to play the whole song back each time you need to
want to test something out.